Refactoring and simplification:
- moved utility functions into utils.generic.py
- simplified the creation of generic metrics
- modified the usage docs to link to the current torchmetrics version docs
