+++ GANDLF/utils/generic.py
@@ -283,3 +283,34 @@ def print_and_format_metrics(
     )
 
     return output_metrics_dict
+
+
+def define_average_type_key(
+    params: Dict[str, Union[Dict[str, Any], Any]], metric_name: str
+):
+    """Determine if the the 'average' filed is defined in the metric config.
+    If not, fallback to the default 'macro'
+    values.
+    Args:
+        params (dict): The parameter dictionary containing training and data information.
+        metric_name (str): The name of the metric.
+    Returns:
+        str: The average type key.
+    """
+    average_type_key = params["metrics"][metric_name].get("average", "macro")
+    return average_type_key
+
+
+def define_multidim_average_type_key(params, metric_name):
+    """Determine if the the 'multidim_average' filed is defined in the metric config.
+    If not, fallback to the default 'global'.
+    Args:
+        params (dict): The parameter dictionary containing training and data information.
+        metric_name (str): The name of the metric.
+    Returns:
+        str: The average type key.
+    """
+    average_type_key = params["metrics"][metric_name].get(
+        "multidim_average", "global"
+    )
+    return average_type_key
